MS1-2 Made arrays for platforms. Draw in loop.
I've added 2 arrays, platform_arr and row_arr. The first element
in each is the bottom platform, and they're sorted in ascending order.

- platform_arr's elements are the leftmost block of each platform
- row_arr's elements are the row indexes of each platform

This commit also moved the drawing of each platform into a loop
that accesses these two new arrays, so we draw whatever the array's
suggest we draw. This means we can control a platform's horizontal
position by simply changing the platform_arr array.