MS1-3 Finalizing MS1. Bug fixes + formatting.
I found a bug in the platform drawing code. The loop incremented
the offset before drawing the first block, so each platform was
shifted to the right by 1 block.

Collision Detection works great, jumping/falling/redrawing map
are all smooth as well. I'm drawing a bit too sporadically to
handle little glitches, so I want to look into that in the next
milestone.